requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
sqlalchemy[asyncio]>=2.0.0

# Pipeline
stageflow-core>=0.3.0

# CLI
click>=8.0.0
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Linting & Type Checking
ruff>=0.1.0
mypy>=1.7.0

# Security
pip-audit>=2.10.0
